+++ backend/main.py
@@ -1,6 +1,8 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from routes import train_routes
+from routes.train_routes import router as train_router
+from routes.structure_routes import router as structure_router
+from routes.control_routes import router as control_router
 
 app = FastAPI(title="SLM Model Builder",
              description="SLM 모델 구조 생성 및 확인 API")
@@ -15,7 +17,9 @@ app.add_middleware(
 )
 
 # 라우터 등록
-app.include_router(train_routes.router, prefix="/api/v1")
+app.include_router(train_router, prefix="/api/v1")
+app.include_router(structure_router, prefix="/api/v1")
+app.include_router(control_router, prefix="/api/v1")
 
 @app.get("/")
 def root():
