+++ backend/ml/models/factory.py
@@ -20,9 +20,10 @@ from .components.dynamic_block import DynamicBlock
 class LayerFactory:
     """메인 팩토리 클래스: 모든 레이어 생성을 담당"""
     @staticmethod
-    def create_layer(node: Dict[str, Any]) -> nn.Module:
+    def create_layer(node: Dict[str, Any], dtype=torch.float32) -> nn.Module:
         layer_type = node["type"]
-        data = node["data"]
+        data = node["data"].copy()
+        data["dtype"] = dtype  # dtype을 data에 추가
         
         factory_map = {
             "tokenEmbedding": TokenEmbeddingFactory,
@@ -31,7 +32,7 @@ class LayerFactory:
             "attention": AttentionFactory,
             "feedForward": FeedForwardFactory,
             "residual": ResidualFactory,
-            "dynamicBlock": DynamicBlockFactory,
+            "transformerBlock": TransformerBlockFactory,
             "linear": LinearFactory,
             "dropout": DropoutFactory
         }
@@ -39,38 +40,39 @@ class LayerFactory:
         factory = factory_map.get(layer_type)
         if factory is None:
             raise ValueError(f"Unknown layer type: {layer_type}")
-            
-        return factory.create(data if layer_type != "dynamicBlock" else node)
+        return factory.create(data if layer_type != "transformerBlock" else node, dtype=dtype)
 
 
 class TokenEmbeddingFactory:
     """토큰 임베딩 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> TokenEmbedding:
+    def create(data: Dict[str, Any], dtype=torch.float32) -> TokenEmbedding:
         return TokenEmbedding(
-            vocab_size=data["vocab_size"],
-            d_model=data["d_model"]
+            vocab_size=data["vocabSize"],
+            d_model=data["embDim"],
+            dtype=dtype  # dtype 추가
         )
 
 
 class PositionalEmbeddingFactory:
     """포지셔널 임베딩 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> Union[
-        LearnedPositionalEmbedding,
-        SinusoidalPositionalEmbedding,
-        RelativePositionalEmbedding,
-        RotaryPositionalEmbedding
-    ]:
+    def create(data: Dict[str, Any], dtype=torch.float32):
         mode = data.get("mode", "learned")
+        common_args = {
+            "ctx_length": data["ctxLength"],
+            "emb_dim": data["embDim"],
+            "dtype": dtype  # dtype 추가
+        }
+        
         if mode == "learned":
-            return LearnedPositionalEmbedding(data["max_len"], data["d_model"])
+            return LearnedPositionalEmbedding(**common_args)
         elif mode == "sinusoidal":
-            return SinusoidalPositionalEmbedding(data["max_len"], data["d_model"])
+            return SinusoidalPositionalEmbedding(**common_args)
         elif mode == "relative":
-            return RelativePositionalEmbedding(data["max_len"], data["d_model"])
+            return RelativePositionalEmbedding(**common_args)
         elif mode == "rotary":
-            return RotaryPositionalEmbedding(data["d_model"], data["max_len"])
+            return RotaryPositionalEmbedding(data["embDim"], data["ctxLength"], dtype=dtype)
         else:
             raise ValueError(f"Unknown positional embedding mode: {mode}")
 
@@ -78,43 +80,37 @@ class PositionalEmbeddingFactory:
 class NormalizationFactory:
     """정규화 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> LayerNorm:
-        return LayerNorm(data["d_model"])
+    def create(data: Dict[str, Any], dtype=torch.float32) -> LayerNorm:
+        d_model = data.get("embDim") or data.get("inDim") or data.get("outDim")
+        return LayerNorm(d_model, dtype=dtype)  # dtype 추가
 
 
 class AttentionFactory:
-    """어텐션 레이어 생성"""
+    """어텐션 레이어 생성 (camelCase 필드에 맞춤)"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> Union[MultiHeadAttentionCombinedQKV, MHAPyTorchScaledDotProduct, GroupedQueryAttention]:
+    def create(data: Dict[str, Any], dtype=torch.float32):
         attn_type = data.get("attn_type", "default")
+
+        common_args = {
+            "d_in": data.get("inDim") or data.get("embDim"),  # inDim 혹은 embDim을 d_in으로 사용
+            "d_out": data.get("outDim") or data.get("embDim"),
+            "num_heads": data["num_heads"],
+            "context_length": data["ctxLength"],
+            "dropout": data.get("dropout", 0.0),
+            "qkv_bias": data.get("qkv_bias", False),
+            "dtype": dtype  # dtype 추가
+        }
+
         if attn_type == "default":
-            return MultiHeadAttentionCombinedQKV(
-                d_in=data["d_in"],
-                d_out=data["d_out"],
-                num_heads=data["num_heads"],
-                context_length=data["context_length"],
-                dropout=data.get("dropout", 0.0),
-                qkv_bias=data.get("qkv_bias", False)
-            )
+            return MultiHeadAttentionCombinedQKV(**common_args)
         elif attn_type == "flash":
-            return MHAPyTorchScaledDotProduct(
-                d_in=data["d_in"],
-                d_out=data["d_out"],
-                num_heads=data["num_heads"],
-                context_length=data["context_length"],
-                dropout=data.get("dropout", 0.0),
-                qkv_bias=data.get("qkv_bias", False)
-            )
+            return MHAPyTorchScaledDotProduct(**common_args)
         elif attn_type == "gqa":
             return GroupedQueryAttention(
-                d_in=data["d_in"],
-                d_out=data["d_out"],
-                context_length=data["context_length"],
-                num_heads=data["num_heads"],
+                **common_args,
                 num_kv_groups=data["num_kv_groups"],
                 rope_base=data.get("rope_base", 10000),
-                rope_config=data.get("rope_config"),
-                dtype=data.get("dtype")
+                rope_config=data.get("rope_config")
             )
         else:
             raise ValueError(f"Unknown attention type: {attn_type}")
@@ -123,31 +119,31 @@ class AttentionFactory:
 class FeedForwardFactory:
     """피드포워드 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> CustomFFN:
+    def create(data: Dict[str, Any], dtype=torch.float32) -> CustomFFN:
         return CustomFFN(
-            emb_dim=data["emb_dim"],
-            dff_ratio=data.get("dff_ratio", 4.0),
+            emb_dim=data.get("embDim") or data.get("inDim"),
+            dff_ratio=data.get("numOfFactor", 4.0),
             activation=data.get("activation", "relu"),
             gated=data.get("gated", False),
-            dtype=torch.float32
+            dtype=dtype  # dtype 이미 있음
         )
 
 
 class ResidualFactory:
     """잔차 연결 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> ResidualConnection:
+    def create(data: Dict[str, Any], dtype=torch.float32) -> ResidualConnection:
         if "source" not in data:
             raise ValueError(f"Residual layer '{data.get('id', 'unknown')}' must have a 'source' field")
         return ResidualConnection(data["source"])
 
 
-class DynamicBlockFactory:
-    """동적 블록 레이어 생성"""
+class TransformerBlockFactory:
+    """트랜스포머 블록 레이어 생성"""
     @staticmethod
-    def create(node: Dict[str, Any]) -> DynamicBlock:
-        children = [LayerFactory.create_layer(child) for child in node.get("children", [])]
-        num_layers = node["data"].get("numLayers", 1)
+    def create(node: Dict[str, Any], dtype=torch.float32) -> DynamicBlock:
+        children = [LayerFactory.create_layer(child, dtype=dtype) for child in node.get("children", [])]
+        num_layers = node["data"].get("numOfBlocks", 1)
         block_id = node["data"].get("id")
         return DynamicBlock(*children, num_layers=num_layers, block_id=block_id)
 
@@ -155,15 +151,19 @@ class DynamicBlockFactory:
 class LinearFactory:
     """선형 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> nn.Linear:
-        return nn.Linear(data["inDim"], data["outDim"])
+    def create(data: Dict[str, Any], dtype=torch.float32) -> nn.Linear:
+        return nn.Linear(
+            in_features=data["inDim"],
+            out_features=data["outDim"],
+            dtype=dtype  # dtype 추가
+        )
 
 
 class DropoutFactory:
     """드롭아웃 레이어 생성"""
     @staticmethod
-    def create(data: Dict[str, Any]) -> nn.Dropout:
-        return nn.Dropout(data.get("dropout_rate", 0.1))
+    def create(data: Dict[str, Any], dtype=torch.float32) -> nn.Dropout:
+        return nn.Dropout(data.get("dropoutRate", 0.1))
 
 
 # ===== Model Classes =====
@@ -210,21 +210,38 @@ class CustomSequential(nn.Module):
 
 
 # ===== Public API =====
-def build_model_from_json(json_list: List[Dict[str, Any]]) -> CustomSequential:
-    """JSON으로부터 모델을 생성하는 메인 함수"""
+def build_model_from_json(json_list: List[Dict[str, Any]], dtype: str = "fp32") -> CustomSequential:
+    """JSON으로부터 모델을 생성하는 메인 함수 (dtype 지원)"""
     id_map = {}
     id_to_module = {}
 
-    for node in json_list:
+    torch_dtype = str_to_torch_dtype(dtype)
+
+    # 첫 번째 객체(config)를 제외한 나머지 레이어들만 처리
+    layer_nodes = [node for node in json_list if "type" in node]
+
+    for node in layer_nodes:
         id_map[node["data"]["id"]] = node
 
     layers = []
-    for node in json_list:
-        layer = LayerFactory.create_layer(node)
+    for node in layer_nodes:
+        layer = LayerFactory.create_layer(node, dtype=torch_dtype)
         if "id" in node["data"]:
             layer.layer_id = node["data"]["id"]
             id_to_module[layer.layer_id] = layer
         layers.append(layer)
 
     return CustomSequential(layers, id_to_module)
+
+
+# 문자열을 torch dtype으로 변환하는 함수 추가
+def str_to_torch_dtype(dtype_str):
+    if dtype_str == "fp32":
+        return torch.float32
+    elif dtype_str == "fp16":
+        return torch.float16
+    elif dtype_str == "bf16":
+        return torch.bfloat16
+    else:
+        raise ValueError(f"Unknown dtype: {dtype_str}")
