+++ backend/tasks/__init__.py
@@ -1,0 +1,2 @@
+from .train import *
+from .structure import *
