+++ backend/temp_structures/e4223abd-eaff-427f-a49f-7a19997ff047.json
@@ -1,0 +1,1 @@
+[{"epochs": 3, "dtype": "fp32"}, {"type": "tokenEmbedding", "data": {"id": "token_emb", "label": "Token Embedding Layer", "vocabSize": 50257, "embDim": 256}}, {"type": "positionalEmbedding", "data": {"id": "pos_emb", "label": "Positional Embedding Layer", "embDim": 256, "ctxLength": 128}}, {"type": "dropout", "data": {"id": "dropout_1", "label": "Dropout Layer", "dropoutRate": 0.1}}, {"type": "transformerBlock", "data": {"id": "block_1", "label": "Transformer Block 1", "numOfBlocks": 2}, "children": [{"type": "normalization", "data": {"id": "norm1", "label": "Norm 1", "inDim": 256}}, {"type": "attention", "data": {"id": "attn1", "label": "Self-Attention", "inDim": 256, "outDim": 256, "num_heads": 4, "ctxLength": 128, "attn_type": "default", "dropout": 0.1, "qkv_bias": false}}, {"type": "dropout", "data": {"id": "dropout_block_1", "label": "Dropout Block 1", "dropoutRate": 0.1}}, {"type": "residual", "data": {"id": "residual_1", "label": "Residual 1", "source": "norm1"}}, {"type": "normalization", "data": {"id": "norm2", "label": "Norm 2", "inDim": 256}}, {"type": "feedForward", "data": {"id": "ffn1", "label": "Feed Forward 1", "numOfFactor": 4, "inDim": 256}}, {"type": "dropout", "data": {"id": "dropout_block_2", "label": "Dropout Block 2", "dropoutRate": 0.1}}, {"type": "residual", "data": {"id": "residual_2", "label": "Residual 2", "source": "ffn1"}}]}, {"type": "normalization", "data": {"id": "final_norm", "label": "Final Norm", "inDim": 256}}, {"type": "linear", "data": {"id": "output_linear", "label": "Output Linear", "inDim": 256, "outDim": 50257}}]
