--- backend/vocab.json
@@ -1,1 +1,0 @@
-{"<pad>": 0, "<unk>": 1, "<|endoftext|>": 2, "a": 3, "about": 4, "again": 5, "am": 6, "amazing": 7, "are": 8, "artificial": 9, "at": 10, "beautiful": 11, "blue": 12, "bright": 13, "bring": 14, "can": 15, "changing": 16, "coffee": 17, "day": 18, "deep": 19, "excited": 20, "feeling": 21, "fine": 22, "for": 23, "future": 24, "go": 25, "going": 26, "good": 27, "goodbye": 28, "grass": 29, "great": 30, "green": 31, "happy": 32, "hello": 33, "how": 34, "i": 35, "i'll": 36, "i'm": 37, "ideahello": 38, "intelligence": 39, "is": 40, "it": 41, "it's": 42, "language": 43, "laptop": 44, "learn": 45, "learning": 46, "let's": 47, "like": 48, "machine": 49, "makes": 50, "me": 51, "meet": 52, "models": 53, "moon": 54, "more": 55, "my": 56, "natural": 57, "networks": 58, "neural": 59, "o'clock": 60, "on": 61, "our": 62, "powerful": 63, "project": 64, "sad": 65, "shop": 66, "sky": 67, "sounds": 68, "store": 69, "sun": 70, "sunny": 71, "that": 72, "that's": 73, "the": 74, "this": 75, "three": 76, "time": 77, "to": 78, "today": 79, "together": 80, "tools": 81, "understand": 82, "walk": 83, "we": 84, "weather": 85, "what": 86, "what's": 87, "where": 88, "work": 89, "world": 90, "would": 91, "yes": 92, "you": 93}
